create /src. moved index.js and defaults to /src
