+++ package.json
@@ -2,7 +2,7 @@
   "name": "cnn-starter-api",
   "version": "1.10.1",
   "description": "Build Graphql APIs in a consistent way",
-  "main": "index.js",
+  "main": "src/index.js",
   "scripts": {
     "test": "node example.js",
     "changelog": "changelog-maker --group"
