+++ src/defaults/schemas/content/data/index.js
