+++ src/defaults/schemas/content/index.js
