+++ src/defaults/schemas/query/index.js
